refactor: restructure actions and services
